requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
httpx==0.25.1
python-dotenv==1.0.0
motor==3.3.1
pymongo==4.6.1
pydantic==2.4.2